+++ src/allpress/exceptions.py
@@ -1,0 +1,9 @@
+class NoParagraphDataError(Exception):
+
+    def __init__(self, msg):
+        super().__init__(msg)
+
+class ForeignKeyWithoutReferenceError(Exception):
+
+    def __init__(self, msg):
+        super().__init__(msg)